runtest: handle FORCE_PAVICS_HOST as work-around for jenkins envinject plugin
envinject is unable to set the 'env' pipeline var so Jenkinsfile was
unable to use it.

But it does set the environment variable properly so build step actually
see it.

This FORCE_PAVICS_HOST is only useful if we want to lockdown the
PAVICS_HOST jenkins build param to a specific host, see usage in
<jenkins-config repo>/jcasc_ouranos/staging_jobs.yaml file.