runtest: change from --nbval to --nbval-lax to not check output by default
For cells that output has to be checked, can add '# NBVAL_CHECK_OUTPUT' to the cell.

The move is primarily due to xarray has enabled html dataset output by
default so strict output checking means we have to revert to the
previous text output that is less useful for the end-users of the
notebook.

We do not want to compromise the documentation role of the
notebooks for the testing role.

There are also other small annoyances like field ordering in the output
or external version numbering change that break the run and require
notebook output update.  These small changes should ideally not force
output update.