runtest: rollback --nbval-lax change in bc4f9daa140d36b86c272ba941cdb9bc33b6401e
Thinking about this over the night.  Maybe disable all output check by default
is probably too extreme.

I would assume there will be at least one cell with `# NBVAL_CHECK_OUTPUT` in
every notebook.  I feel that output checking is the equivalent of the `assert`
in regular unit tests so it would be weird and a bit scary that we do not
assert anything in the notebooks.

So either have `# NBVAL_CHECK_OUTPUT` with `--nbval-lax` or `#
NBVAL_IGNORE_OUTPUT` with regular `--nbval` is the same: we can never hide that
we leverage the notebooks as test cases.  Might as well be on the safe side and
stick with regular `--nbval` like currently.

Most of the current cells do not need  `# NBVAL_IGNORE_OUTPUT`.  We will need
it for cells that output xarray dataset in html.  So probably 1 or 2 cells per
notebooks, assuming notebooks are not gigantic.

For other ordering problems in the output, we can also use `#
NBVAL_IGNORE_OUTPUT` instead of doing hacks like in
https://github.com/Ouranosinc/pavics-sdi/commit/d19ad355b522b6f5f012308b4fe94a079ca76c37
which makes the code harder to read from the tutorial perspective.