pavics-landing: delegate required creation of Jupyter env layout
In addition to the writable-workspace/ folder, we also need to re-create
the pavics-homepage/ layout.

To fix
```
DriverError: /notebook_dir/pavics-homepage/tutorial_data/test_regions.geojson: No such file or directory
```

Path to those .geojson files are hardcoded so users can copy the nb to
writable-workspace/ dir and still be able to run them seemlessly from
the Jupyter env (without having to also copy those *.geojson files with
the notebooks).

The layout creation will be delegated to pavics-landing repo so this
repo is generic.