docker: replace jupyter-dash with dash since the previous have warnings, keep plotly
Previous jupyter-dash warning "Plugin 'jupyterlab_dash' is already
registered."

Plotly will be pulled as indirect dependency.

Related issues and discussions:

https://github.com/plotly/jupyter-dash/issues/49#issuecomment-1992606506
This entire package is obsolete, its functionality has been integrated
into the main Dash package. If anyone is interested in helping modernize
JupyterDash packaging the Dash repo is the place to continue this
discussion.

https://github.com/Ouranosinc/pavics-jupyter-env-issues/issues/5#issuecomment-1994587587
@Zeitsperre
As far as I can tell, we abandoned Dash in favour of Holoviews. Fell
free to drop it.

https://github.com/Ouranosinc/pavics-jupyter-env-issues/issues/5#issuecomment-1994689326
@tlogan2000
Dash/jupyter dash can be dropped I think but `plotly` itself can be used
as a backend in Panel/Hvplot/Holoviews ecosystem so would be good to
keep.