docker: add conda-pack to root conda env
To allow using this exact Jupyter env, outside of Jupyter, on some local machine.

https://conda.github.io/conda-pack/

conda-pack is a command line tool for creating archives of conda environments
that can be installed on other systems and locations. This is useful for
deploying code in a consistent environment—potentially where python and/or
conda isn’t already installed.

A tool like conda-pack is necessary because conda environments are not
relocatable. Simply moving an environment to a different directory can render
it partially or completely inoperable. conda-pack addresses this challenge by
building archives from original conda package sources and reproducing conda’s
own relocation logic.